Add files via upload
In this project, we aim to create a game using Docker and deploy it to AWS cloud. Docker is a containerization technology that allows developers to package and distribute their applications in a consistent and portable manner. AWS cloud provides a scalable, reliable, and secure platform for deploying applications.

We first designed and developed the game, ensuring it was compatible with Docker containers. We then used Docker to package the game and its dependencies into a container image. We also set up an AWS EC2 instance and configured it to host the Docker container. Finally, we deployed the Docker container to the AWS cloud and verified that the game was running correctly.

Throughout the project, we encountered several challenges, including configuring network settings, optimizing resource utilization, and debugging issues with the container image. We discuss these challenges in detail and provide solutions for each of them.

The result of the project is a functional game that can be easily deployed and scaled using Docker containers on AWS cloud. This project report serves as a guide for developers interested in creating games using Docker and deploying them to AWS cloud.